+++ buildpacks/erlang-otp/Cargo.toml
@@ -14,6 +14,7 @@ tar = { version = "0.4", default-features = false }
 flate2 = { version = "1", default-features = false, features = ["zlib"] }
 
 [dev-dependencies]
+buildpack-test-support = { path = "../../test-support" }
 libcnb-test = "=0.22.0"
 serde_json = "1"
 tempfile = "3"
